database_updater_function/requirements.txt
# Cloud Function dependencies
requests>=2.25.0
google-cloud-storage>=2.0.0
functions-framework>=3.0.0
